+++ CHANGELOG.md
@@ -42,6 +42,8 @@
   - `forest-cli db clean` to `forest-tool db destroy`
 - [#3362](https://github.com/ChainSafe/forest/pull/3362) Moved the following
   `forest-cli wallet` subcommands to `forest-wallet`
+- [#3435](https://github.com/ChainSafe/forest/pull/3435) Moved subcommand
+  `forest-cli car concat` subcommands to `forest-tool`
 
 ### Added